- Threading now works!!!
You can submit a commentary on any paper, or a later version of any of your
own papers. If it's accepted, the version and commentary threads of all
affected papers are updated.

Likewise, if a paper is withdrawn, database integrity is maintained; all
references to the paper are removed and appropriate threads updated.

The only part that isn't implemented is the highlighting of any
`certified' version.